requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
SQLAlchemy==2.0.21
pydantic~=2.4.2
cryptography~=41.0.4
flask~=3.0.0